requirements.txt
langgraph>=0.2.21
pytrends>=4.9.2
beautifulsoup4>=4.12.3
pandas>=2.2.2
requests>=2.32.3
llama-cpp-python>=0.2.84
python-dotenv>=1.0.1
fastapi>=0.115.5
uvicorn>=0.32.1
